Added Game Class, Defined Player and User Class Better
Added a game class, defined the teh user and AI classes better
New code successfully integrated and running
