+++ LazyAI.java
@@ -7,11 +7,11 @@ public class LazyAI extends Players {
 		super();
 	}
 
-	public char charGuess() {
+	public char getInput() {
 		char [] alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ".toCharArray();
 		int rand = (int)(Math.random() * (26-1)) + 1;
 		char g = alphabet[rand - 1];
-		
+		setCurrGuess(g);
 		return g;
 	}
